Disable TestValidOpenAPISpec when -race test flag exists
